+++ backend/models/users.go
@@ -21,15 +21,4 @@ type Error struct {
 	Message string `json:"Message"`
 }
 
-type Transaction struct {
-	Amount	    float32		 `json:"amount"`
-	Name		    string `json:"name"`
-	Date        string `json:"date"`
-	Category    string `json:"category"`
-	Description string `json:"description"`
-}
 
-type FinancialInfo struct {
-	TransactionID	 uint				 `json:"ID"`
-	Transactions []Transaction `json:"transactions" gorm:"foreignKey:Transactions"`
-}
